test.py
import torch, numpy, yaml
print("torch", torch.__version__)
print("numpy", numpy.__version__)
print("cuda_available", torch.cuda.is_available())
print("cuda_device_count", torch.cuda.device_count())
if torch.cuda.is_available():
    print("device0", torch.cuda.get_device_name(0))